prepare for IDF5.1

+++ lib/rtime/rtime.cpp
@@ -270,7 +270,7 @@ boolean RTIME::obtain_time(){
     int32_t retry = 0;
     const int32_t retry_count = 10;
     while(timeinfo.tm_year < (2016 - 1900) && ++retry < retry_count) {
-        sprintf(sbuf, "Waiting for system time to be set... (%d/%d)", retry, retry_count);
+        sprintf(sbuf, "Waiting for system time to be set... (%ld/%ld)", (long int)retry, (long int)retry_count);
         if (RTIME_info) RTIME_info(sbuf);
         vTaskDelay(uint16_t(2000 / portTICK_PERIOD_MS));
         time(&now);
